config is no longer a global but instead references the instances config

+++ lib/spontaneous/plugins/publishing.rb
@@ -163,6 +163,7 @@ module Spontaneous::Plugins
       end
 
       def delete_revision(revision)
+        return if revision.nil?
         database.drop_table(revision_table(revision)) if revision_exists?(revision)
       end
